Use a QMap to simplify CasterBoard
Instead of using separate member variables for every player, a single
map can both store the players and provide a convenient  mechanism for
mapping key press events to the appropriate player.

The map uses a Qt::Key symbol as the key and stores the pointer to each
widget as the value.

To populate the map, we iterate over each character in a string, where
the characters correspond to the labels of the hotkey widgets. The
corresponding Qt::Key values are stored in another array that is
accessed in parallel. A counter keeps track of the offset in the keys
array and is also used to calculate the grid position of each widget.